refactor ticket detail view with styles
